[FEATURE][processing] Add "convert to curves" algorithm
Converts a linear geometry type to the corresponding curved
geometry type, attemping to identify segments in the original
geometries which can be replaced by arcs.

The distance tolerance parameter specifies the maximum distance
allowed between the original location of vertices and where they
would fall on the converted curved geometries.

This algorithm only consider a segments as suitable for
replacing with an arc if the points are all regularly spaced on
the candidate arc. The angle tolerance parameter specifies the
maximum angular deviation (in degrees) allowed when testing for
regular point spacing.

Already curved geometries will be retained without change.